Remove utils dir
