+++ sui/src/lib.rs
@@ -5,5 +5,4 @@
 pub mod config;
 pub mod shell;
 pub mod sui_commands;
-pub mod utils;
 pub mod wallet_commands;
