+++ sui/src/wallet.rs
@@ -6,9 +6,8 @@ use std::io;
 use std::path::PathBuf;
 use structopt::clap::{App, AppSettings};
 use structopt::StructOpt;
-use sui::config::WalletConfig;
+use sui::config::{Config, WalletConfig};
 use sui::shell::{AsyncHandler, CommandStructure, Shell};
-use sui::utils::Config;
 use sui::wallet_commands::*;
 
 const FAST_X: &str = "   _____       _    _       __      ____     __
